GenAI_Code/requirements.txt
flask
gunicorn
python-dotenv
logging
groq
requests
beautifulsoup4